requirements.txt
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.29.4
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msrest==0.7.1
oauthlib==3.3.1
packaging==25.0
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
